SICA - Estandarización general en personas y asistencias a eventos

+++ Modules/SICA/Database/Migrations/2020_09_07_000007_create_people_table.php
@@ -65,9 +65,9 @@ class CreatePeopleTable extends Migration
             ])->nullable();
             $table->enum('sisben_level', ['A','B','C','D'])->nullable();
             $table->string('address')->nullable();
-            $table->unsignedInteger('telephone1')->nullable();
-            $table->unsignedInteger('telephone2')->nullable();
-            $table->unsignedInteger('telephone3')->nullable();
+            $table->unsignedBigInteger('telephone1')->nullable();
+            $table->unsignedBigInteger('telephone2')->nullable();
+            $table->unsignedBigInteger('telephone3')->nullable();
             $table->string('personal_email')->nullable();
             $table->string('misena_email')->nullable();
             $table->string('sena_email')->nullable();
